requirements.txt
pandas==2.1.4
python-dotenv==1.0.0
azure-storage-blob==12.19.0